src/requirements.txt
nltk==3.8.1
noisereduce==2.0.1
pymodm==0.4.3
pymongo==3.11.3
pymystem3==0.2.0
python-telegram-bot==13.4.1
pytz==2024.1
Requests==2.30.0
schedule==1.1.0
scipy==1.13.0
python-gettext==5.0
pip==23.1.2
